Fix: Use stream deletion function when cleaning up
In theory, once the destroy stream ht function is called with the hash
table, it should be empty. However, for some fatal errors, it might not
so it's imperative that we gracefully delete the stream and free it
using an RCU call so both hash tables (stream and the one for the
pending command) are synchronized.

Simply freeing the stream could have created possible fd leaks and
invalid node for the data pending hash table.

Signed-off-by: David Goulet <dgoulet@efficios.com>