Fix: set relayd sock sent flag per consumer socket
The session daemon usually handles two consumers, one for 32 bits
applications and one for 64 bits.

When sending relayd information to the consumer, we must sent it to
every existing consumer in case a session contains 32 and 64 bits
applications. So, the sent flag is moved from the consumer output object
to be per consumer socket.

This bug was seen and diagnosed during the 2.2 development phase and is
fixed in master branch in the following commit.

commit (master): ffe600149a7608221985751e1bf293234bf2545c

Signed-off-by: David Goulet <dgoulet@efficios.com>