Fix: put session list lock around the app registration
A successful call from ust_app_register makes the application visible to
the tracing registry on the client side so by moving this lock at the
start of the registration process until the end, we are certain that the
application does not receive an out of order message during
registration.

It basically protects the application registration from tracing command
such as enable-event.

This has been fixed in the master branch by an other commit during the
2.2 development phase.

Signed-off-by: David Goulet <dgoulet@efficios.com>