Fix: don't start the relayd with a wrong --output dir
The -o, --output option never supported a relative path but still the
relayd was able to start.

This makes sure the relayd does not start without an absolute pathname.
The help and man are also updated to document this behavior.

Ticket #523 has been opened as a feature request to support relative
path.

Fixes #481

Acked-by: Julien Desfossez <julien.desfossez@efficios.com>
Signed-off-by: David Goulet <dgoulet@efficios.com>