Fix: send per-pid session id in channel creation
The registry indexing for per-pid sessions is done with a per-pid
session id. So for per-pid buffers, we need to send the per-pid session
id as well as the global session id to the consumer in order to give it
enough information if it needs to request metadata later.

This patch adds the session_id_per_pid to the channel creation message
and to the consumer. When the sessiond receives a metadata_request,
depending on the buffer type (per-pid or per-uid), it selects the right
id to do the registry lookup.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>
Signed-off-by: David Goulet <dgoulet@efficios.com>

Conflicts:
	src/bin/lttng-sessiond/consumer.c
	src/bin/lttng-sessiond/consumer.h
	src/bin/lttng-sessiond/ust-consumer.c
	src/common/consumer.c
	src/common/consumer.h
	src/common/kernel-consumer/kernel-consumer.c
	src/common/sessiond-comm/sessiond-comm.h
	src/common/ust-consumer/ust-consumer.c